tweak: fix some warnings

+++ Code/connect/include/Packet.hpp
@@ -10,7 +10,7 @@ namespace TiltedPhoques
         TP_ALLOCATOR;
 
         Packet() noexcept;
-        Packet(size_t aSize) noexcept;
+        Packet(uint32_t aSize) noexcept;
         virtual ~Packet() noexcept;
 
         TP_NOCOPYMOVE(Packet);
@@ -19,10 +19,10 @@ namespace TiltedPhoques
         [[nodiscard]] char* GetData() const noexcept;
 
         // Get the writable data size, note that the actual packet size may differ from that
-        [[nodiscard]] size_t GetSize() const noexcept;
+        [[nodiscard]] uint32_t GetSize() const noexcept;
 
         // Get the writable data size + protocol data size
-        [[nodiscard]] size_t GetTotalSize() const noexcept;
+        [[nodiscard]] uint32_t GetTotalSize() const noexcept;
 
         // Returns true if the packet has an associated buffer, usually used to check if the underlying allocator had enough space
         [[nodiscard]] bool IsValid() const noexcept;
@@ -33,12 +33,12 @@ namespace TiltedPhoques
         friend struct Client;
 
         char* m_pData;
-        size_t m_size;
+        uint32_t m_size;
     };
 
     struct PacketView : Packet
     {
-        PacketView(char* aPointer, size_t aSize);
+        PacketView(char* aPointer, uint32_t aSize);
         virtual ~PacketView();
     };
 }
